Move account form to its own component file
